pyproject.toml
[tool.poetry]
name = "PipeX"
version = "0.2.0"
description = "A powerful CLI-based ETL (Extract, Transform, Load) pipeline automation tool for data engineering workflows"
authors = ["ImCYMBIOT <agniveshkumar15@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/ImCYMBIOT/pipex"
repository = "https://github.com/ImCYMBIOT/pipex"
keywords = ["etl", "data-pipeline", "cli", "automation", "data-engineering"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Database",
    "Topic :: System :: Archiving",
]

[tool.poetry.scripts]
pipex = "app.cli:app"

[tool.poetry.dependencies]
python = "^3.11"
# Core CLI and data processing
typer = "^0.15.0"
pandas = "^2.2.0"
pyyaml = "^6.0.0"
python-dotenv = "^1.0.0"
tqdm = "^4.67.0"

# HTTP and API clients
requests = "^2.32.0"
cachetools = "^5.5.0"

# Database connectors
SQLAlchemy = "^2.0.0"
mysql-connector-python = "^9.1.0"
psycopg2 = "^2.9.0"
pymongo = "^4.10.0"

# Data validation
jsonschema = "^4.23.0"

# File format support
openpyxl = { version = "^3.1.0", optional = true }
xlrd = { version = "^2.0.0", optional = true }
lxml = { version = "^5.0.0", optional = true }
pyarrow = { version = "^17.0.0", optional = true }

# Cloud storage (optional)
boto3 = { version = "^1.35.0", optional = true }
google-cloud-storage = { version = "^2.10.0", optional = true }
azure-storage-blob = { version = "^12.19.0", optional = true }

[tool.poetry.extras]
aws = ["boto3"]
gcp = ["google-cloud-storage"]
azure = ["azure-storage-blob"]
excel = ["openpyxl", "xlrd"]
xml = ["lxml"]
parquet = ["pyarrow"]
all = [
    "boto3", 
    "google-cloud-storage", 
    "azure-storage-blob", 
    "openpyxl", 
    "xlrd", 
    "lxml", 
    "pyarrow"
]

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^8.3.0"
pytest-mock = "^3.14.0"
pytest-cov = "^5.0.0"

# Code quality
black = "^24.10.0"
flake8 = "^7.1.0"
mypy = "^1.13.0"
isort = "^5.13.0"

# Development tools
pre-commit = "^4.0.0"

[tool.poetry.group.optional.dependencies]
# Optional dependencies for specific features
questionary = "^2.0.0"  # Interactive prompts
rich = "^13.9.0"        # Rich terminal output

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--verbose",
    "--cov=app",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=80"
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__init__.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]
